backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.1.0.post1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
firebase-admin==6.4.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart>=0.0.9
email-validator==2.1.0.post1
resend==2.13.0